backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Google APIs
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0


# RAG & Embeddings (sans torch - il s'installe automatiquement)
chromadb==0.4.24
sentence-transformers==2.2.2

# Data Processing
numpy==1.24.3

# Validation & Utils
pydantic==2.5.0
requests==2.31.0

# Testing
pytest==7.4.3
jupyter==1.0.0

# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# LLM & AI (sans LangChain pour l'instant)
google-generativeai==0.5.0

# HTTP & Data
requests==2.31.0
pydantic==2.5.0
